feat: StatementStrategy 익명 클래스를 이용해 생성
